Add trivial search
The query returning `Book` was throwing an exception when it found
nothing.  So I switched to `List<Book>` to:

1. potentially match multiple book titles
2. return an empty list when nothing is found

The documentation seemed to indicate that a simple `find(title)` query
was "less flexible", which I took to mean it is less of a best practice.
So, I switched to `findByTitleIlike(title)` as something more explicit,
and I got case-insensitive partial searches for free.

See [Micronaut Data documentation](https://micronaut-projects.github.io/micronaut-data/latest/guide/#querying).